[BREAKING] Rename skinTex and skinBatchTex chunk names to skin and skinBatch
