+++ src/scene/shader-lib/chunks/common/vert/skin.js
