+++ src/scene/shader-lib/chunks/common/vert/skinBatch.js
